+++ MainProject/Injecter/IScopeStore.cs
@@ -4,8 +4,8 @@ namespace Injecter
 {
     public interface IScopeStore
     {
-        IServiceScope CreateScope(object owner);
-        void DisposeScope(object owner);
-        IServiceScope GetScope(object owner);
+        IServiceScope CreateScope<T>(T owner) where T : notnull;
+        void DisposeScope<T>(T owner) where T : notnull;
+        IServiceScope? GetScope<T>(T owner) where T : notnull;
     }
 }