+++ UnityProject/Injecter.Unity/Assets/Injecter.Unity/InjectedMonoBehaviorT.cs
@@ -1,14 +1,15 @@
-﻿using System;
+﻿#nullable enable
+using System;
 
 namespace Injecter.Unity
 {
-    public abstract class InjectedMonoBehavior<TViewModel> : InjectedMonoBehavior
+    public abstract class InjectedMonoBehaviour<TViewModel> : InjectedMonoBehaviour
     {
-        protected InjectedMonoBehavior(bool createScopes) : base(createScopes)
+        protected InjectedMonoBehaviour(bool createScopes) : base(createScopes)
         {
         }
 
-        [Inject] protected TViewModel ViewModel { get; } = default;
+        [Inject] protected TViewModel ViewModel { get; } = default!;
 
         protected override void OnDestroy()
         {