+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # 7.0.0
+- Renamed classes: \*MonoBehavior\* -> \*MonoBehavio**u**r\*
+- InjectedMonoBehavior now removes Scope from the store during `OnDestroy`
+- ScopeStore is now generic
 - Update dependencies
 - Update minimum Unity version to `2020.3`
 - Update to .NET 6 where not breaking